+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Change History
 
 ## master
+- Added command repeat information to status bar
 
 ## 0.3.1
 - Added `g` (top of list) and `G` bottom of list motions ([#5](https://github.com/LPGhatguy/magic-school-bus/issues/5))